requirements.txt
flask
openai
gTTS
requests
python-dotenv
SpeechRecognition